requirements.txt
tensorflow==2.15.0
keras==2.15.0
streamlit==1.28.1
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
scikit-learn==1.3.0
pillow==10.0.0
tqdm==4.66.1 